♻️Refactor code
* Rename editNRemove.js to crud.js
* Create helpers folder and moved helpers.js there.
* Create modules folder and moved crud.js there
* Move some logic from index.js to crud.js
