+++ .babelrc
@@ -1,5 +1,7 @@
 {
   "presets": [
-    ["@babel/preset-env"]
+    [
+      "@babel/preset-env"
+    ]
   ]
 }
