+++ modules/crud.js
@@ -1,3 +1,21 @@
+import { nanoid } from 'nanoid';
+
+const addTask = (event, tasks) => {
+  tasks = JSON.parse(localStorage.getItem('tasks')) || [];
+  event.preventDefault();
+  const task = {
+    description: event.target[0].value,
+    completed: false,
+    id: nanoid(5),
+  };
+  if (task.description === '') {
+    return;
+  }
+  tasks.push(task);
+  localStorage.setItem('tasks', JSON.stringify(tasks));
+  event.target[0].value = '';
+};
+
 const editTask = (e, tasks) => {
   const taskId = e.target.parentNode.parentNode.id;
   const taskIndex = tasks.findIndex((task) => task.id === taskId);
@@ -13,4 +31,4 @@ const removeTask = (e, tasks) => {
   localStorage.setItem('tasks', JSON.stringify(tasks));
 };
 
-export { editTask, removeTask };
+export { addTask, editTask, removeTask };
